Add two test cases that covers clearGroup and observeDestinationChanges.
One of them adds 4 fragments, and calls clearGroup to remove last two fragment.

Other one, the one fails, adds four fragments and clearGroup to remove middle two fragments. The callback we passed into the observeDestinationChanges should not be triggered because of this operation because the actual destination has not changed because of the clearGroup operation. User is still seeing the same page.